+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-05-06
+
+### Changed
+
+- Improve coach logging.
+- Use repository images for bot comments and notifications.
+
+### Fixed
+
+- Dojo coach wasn't able to merge PR.
+
 ## [1.1.0] - 2020-04-08
 
 ### Added